requirements.txt
streamlit>=1.36.0
openai>=1.40.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
langchain>=0.2.10
langchain-openai>=0.1.17
faiss-cpu>=1.8.0
pymupdf>=1.24.9
docx2txt>=0.8
